review: address PR #73 inline comments
Three fixes suggested in the PR review:

1. ``pawn/shard_loader.py``: warn when a cached shard can't be counted.
   Previously a corrupt or unreadable shard silently contributed 0
   rows to the budget pre-count, causing the main download loop to
   overshoot ``max_games``. Now we print a warning so the discrepancy
   is visible.

2. ``pawn/adapter_training.py``: use ``dtype=torch.int64`` for the
   train top-1 accumulators. ``(preds == targets).sum()`` returns
   int64; the float32 accumulators were silently casting on ``+=``.
   Making the dtype explicit matches the semantic (counting correct
   predictions) and avoids a potential dtype-promotion edge case.

3. ``tests/lab/test_runner.py``: switch the ``asyncio.sleep`` patch
   from ``MagicMock(return_value=asyncio.sleep(0))`` to
   ``new_callable=AsyncMock``. The old mock yielded a single coroutine
   at construction time, which can only be awaited once. The tests
   happened to pass because ``is_alive`` fires on the first loop
   iteration, but AsyncMock is the correct tool and doesn't rely on
   that accident.

All 1,427 tests still pass.

+++ pawn/adapter_training.py
@@ -959,10 +959,13 @@ def train(
         # while the CPU reads the scalar for batch N — on ROCm this shows
         # up as ~30% GPU utilization even when the data pipeline is idle.
         epoch_loss_t = torch.zeros((), device=device)
-        epoch_top1_t = torch.zeros((), device=device)
+        # top-1 accumulators hold a count of correct predictions, so
+        # they're int64 to match ``(preds == targets).sum()``. Keeping
+        # the dtype explicit avoids a silent float32 cast on ``+=``.
+        epoch_top1_t = torch.zeros((), device=device, dtype=torch.int64)
         epoch_positions = 0
         log_loss_t = torch.zeros((), device=device)
-        log_top1_t = torch.zeros((), device=device)
+        log_top1_t = torch.zeros((), device=device, dtype=torch.int64)
         log_positions = 0
         t0 = time.time()
